requirements.txt
redis
flask
flask_cors
web3==5.31.3
waitress
python-dotenv
numpy